Chenged inmemory search with Lucene.NET

+++ Doco/Core/Models/SearchResult.cs
@@ -3,7 +3,7 @@
     public class SearchResult
     {
         public string DocumentPath { get; set; } = string.Empty;
-        public int Score { get; set; }
+        public float Score { get; set; }
         public DocumentMetadata Metadata { get; set; } = new DocumentMetadata();
     }
